+++ internal/config/v3config/config_test.go
@@ -269,7 +269,7 @@ assets:
       image: assets/dashboard/simple/panel/background.png
 logs:
   jsonl:
-    path: logs/godrivelog.jsonl
+    path: logs/gogauges.jsonl
     sensors:
       - speed
 dashboards: