+++ internal/dashboard/v3dashboard/dashboard.go
@@ -8,10 +8,10 @@ import (
 	"strings"
 	"time"
 
-	v3assets "github.com/MickMake/GoDriveLog/internal/assets"
-	"github.com/MickMake/GoDriveLog/internal/config/v3config"
-	v3gauges "github.com/MickMake/GoDriveLog/internal/dashboard/gauges"
-	"github.com/MickMake/GoDriveLog/internal/sensors"
+	v3assets "github.com/MickMake/GoGauges/internal/assets"
+	"github.com/MickMake/GoGauges/internal/config/v3config"
+	v3gauges "github.com/MickMake/GoGauges/internal/dashboard/gauges"
+	"github.com/MickMake/GoGauges/internal/sensors"
 )
 
 var gaugePackageLoader = v3gauges.LoadPackageWithSearchPaths